Split plugins into modules; Add keyvalue_dictionary
