+++ lib/mapp.mjs
@@ -24,7 +24,7 @@ import location from './location/_location.mjs'
 
 import Mapview from './mapview/_mapview.mjs'
 
-import * as plugins from './plugins.mjs'
+import plugins from './plugins/_plugins.mjs'
 
 hooks.parse();
